Upgrade strscan tests to integration test format
See GH-628.
