+++ artichoke-backend/src/extn/stdlib/strscan/strscan.rb
